BufferQueue backed C2BlockPool
IGBP based BlockPool is implemented. Without IGBP the BlockPool behaves
exactly same as C2BasicGraphicBlockPool.

After IGBP is registered, the BlockPool gives a block based on graphic
buffer of the IGBP. IGBP id and slot id is given via ExtraData of
C2Handle. The information can be used to decide whether to attach the
buffer or not.

Bug: 74077399
Change-Id: Id12f183b8e7853bfd993588016f8f35faaca716c