DO NOT MERGE: Remove unused `tags` property from Android.bp files
The tags property is (and has always been) unused by Soong. The property has
been defined as a list of strings, and the `androidmk` converted any
LOCAL_MODULE_TAGS entries over to it, but we've never done anything with it.

In preparation for removing the definition from Soong, I'm removing it from all
Android.bp files in the tree.

Since this has never done anything, this is a no-op, but if you really did want
the Android.mk behavior, the proper way to define a module to be installed in
userdebug / eng builds is to use PRODUCT_PACKAGES_DEBUG or PRODUCT_PACKAGES_ENG
in the appropriate product makefile.

Bug: 80085975
Change-Id: If83e311dcafd62722689b0406e7ec85793dcfab0
Merged-In: If83e311dcafd62722689b0406e7ec85793dcfab0
Exempt-From-Owner-Approval: global no-op build change
Test: remove `tags` from Soong, see errors go away.