CCodec: query in/out allocator ID and create remote output block pool
The flow of CCodec to specify input and output allocators and block pools:

1. On start CCodec should query C2PortAllocatorsTuning::input and ::output from
component.

2. If C2PortAllocatorsTuning::input is valid, CCodec should create local block
pool of input buffers with its value as allocator ID. If is invalid, CCodec
creates with default allocator ID.

3. If C2PortAllocatorsTuning::output is valid, CCodec should create remote
block pool of output buffers with its value as allocator ID (and obtained the
block pool ID). If is invalid, CCodec does nothing.

4. If remote output block pool is created, try to set output surface.

5. CCodec should config C2PortBlockPoolsTuning::output to component with
obtained block pool ID. If this parameter is not supported by component,
component will creates the default local block pool on its own.

Bug: 79392858
Test: screenrecord and stagefright playback
Test: CtsMediaTestCases android.media.cts.MediaPlayerTest#testLocalVideo_MP4_H264_480x360_500kbps_30fps_AAC_Stereo_128kbps_44110Hz
Change-Id: I39271373e7a581dee85536ac25a79de3a7a5f36f