C2: protect C2InterfaceHelper with a mutex
It is more straightforward to add a mutex here and expose a getter
(lock()) than to add it to the wrappers where it is used, as it is
subclassed and wrapped to expose C2ComponentInterface. SimpleC2Component
only has access to either the derived class from C2InterfaceHelper or
to the wrapped C2ComponentInterface.

Bug: 12345678
Change-Id: I66ff27a5b0db5bb5e22b4e3d7523ca03d3beb881