CCodec: use bufferchannel for setParams only if using bytebuffer input
Input surface mode does not go through the bufferchannel. Moreover, the synchronization
between setParam and queue is not defined for input surface mode. (E.g. is it synchronized
to the next frame queued to the component, the next acquired frame from the surface, or
the next frame queued to the input surface? - the latter would correspond most closely to
the bytebuffer interpretation).

Without this fix, setParameters do not take effect for input surface mode.

Bug: 67591367
Test: sanity test
Change-Id: Iebc21b2eeee40dfdd10dcc1cd425902d182cae18