Fix adaptive playback in HEVC decoder
In case of adaptive playback in HEVC, where header and first frame of
new resolution were sent to the decoder after change in resolution,
decoder was only decoding header and was ignoring the frame.
This was because the plugin did not have any info on updated dimensions
and output buffer was still allocated as per older resolution.
This resulted in visual artifacts while running following CTS

As AVC decoder also has similar API, same changes are done
there as well.

Test: cts-tradefed run commandAndExit cts-dev -m CtsMediaTestCases \
-t android.media.cts.AdaptivePlaybackTest#testHEVC_adaptiveDrc
Bug: 112776299

Change-Id: Iaea8d777f8fefd5f4b48f894e6115374f4fd7a95