Handle first input with header and picture sent for all i cases
If first input to HEVC/AVC c2 decoder contains picture and header
together and if it is an all intra sequence with zero display delay,
then decoders were decoding first frame as well.
Since output width and height is not updated, C2 plugin expects that
first call doesn't produce an output.
This is fixed now, by setting the decoder in header decode mode first
and then switching to frame decode mode after an header is decoded.

Bug: 113296879
Test: cts-tradefed run commandAndExit cts-dev -m CtsMediaTestCases\
 -t android.media.cts.DecoderTest
Test: cts-tradefed run commandAndExit cts-dev -m CtsMediaTestCases\
-t android.media.cts.AdaptivePlaybackTest

Change-Id: Ia377ad01e40e5d3ba575f8da1471f55acd023fc6