codec2: do setParameters on MediaCodec's looper
Currently, when CCodec is used, the setParameter is
performed on CCodec's looper, while queueInputBuffer
happens on MediaCodec's looper. For video codecs, the
two loopers are different. This could make setParameters
and the input buffer out of sync.

Calling setParameters() directly in
CCodec::signalSetParameters(), so that it happens on
MediaCodec's looper and sync'ed with queueInputBuffers.

bug: 118507186
test: test locally by calling MediaCodec.setParameters
with newly added HDR10+ metadata, and forcing C2SoftVpxDec
to be used.
Change-Id: I10bad80edc7ecda4bc4df820bdbcfa10b9b841be