CCodecBufferChannel: Add the allocated block to ref-tracker in EncryptedLinearInputBuffers
When a EncryptedLinearBlockBuffer is allocated, update the
weak-reference to that block in the corresponding slot so
it does not get overwritten by subsequent allocations.

Fixes corruption issue with secure video playback (esp. for apps
that do not call MediaCodec.getInputBuffer() for every input available;
but rely on {index, ByteBuffer} association from an initial call to
MediaCodec.getInputBuffers(..))

Bug: 117861662
Test: play video on Netflix player, Youtube, Local playback

Change-Id: I9200438a552a20cfabe0ee5f39e6d56729205ae9